+++ tests/covalent_tests/dispatcher_plugins/local_test.py
@@ -23,7 +23,57 @@
 
 import pytest
 
-from covalent._dispatcher_plugins.local import LocalDispatcher
+import covalent as ct
+from covalent._dispatcher_plugins.local import LocalDispatcher, get_redispatch_request_body
+
+
+def test_get_redispatch_request_body_null_arguments():
+    """Test the get request body function with null arguments."""
+
+    @ct.electron
+    def identity(a):
+        return a
+
+    @ct.electron
+    def add(a, b):
+        return a + b
+
+    response = get_redispatch_request_body(
+        "mock-dispatch-id",
+    )
+    assert response == {
+        "json_lattice": None,
+        "dispatch_id": "mock-dispatch-id",
+        "electron_updates": {},
+        "reuse_previous_results": False,
+    }
+
+
+def test_get_request_body_args_kwargs(mocker):
+    """Test the get request body function when args/kwargs is not null."""
+    generate_electron_updates_mock = mocker.patch(
+        "covalent._dispatcher_plugins.redispatch_utils._generate_electron_updates",
+        return_value="mock-electron-updates",
+    )
+    get_result_mock = mocker.patch("covalent._dispatcher_plugins.redispatch_utils.get_result")
+    get_result_mock().lattice.serialize_to_json.return_value = "mock-json-lattice"
+
+    response = get_redispatch_request_body(
+        "mock-dispatch-id",
+        new_args=[1, 2],
+        new_kwargs={"a": 1, "b": 2},
+        replace_electrons={"mock-task-id": "mock-electron"},
+    )
+    generate_electron_updates_mock.assert_called_once_with(
+        "mock-dispatch-id", {"mock-task-id": "mock-electron"}
+    )
+    assert response == {
+        "json_lattice": "mock-json-lattice",
+        "dispatch_id": "mock-dispatch-id",
+        "electron_updates": "mock-electron-updates",
+        "reuse_previous_results": False,
+    }
+    get_result_mock().lattice.build_graph.assert_called_once_with(*[1, 2], **{"a": 1, "b": 2})
 
 
 @pytest.mark.parametrize(
